Remove dtype attribute from BaseWavefunction
This attribute existed for historical reasons when we wanted to do some fancier
things with the data types (do different solver behaviours for complex
parameters). Because the solvers of wavefunctions used to be inside the
wavefunction as a method, this attribute was needed to control the solver
behaviour. However, with the current structure of the code, we can just let
the solver behave differently based on the data types of the parameters. I doubt
that the parameters and the overlap will have different data types

1. Remove assign_dtype method and dtype attribute from BaseWavefunction
2. Add dtype parameter
3. Make corresponding changes to all inherite __init__, docstrings, and tests