Move memory management to objective class
The main purpose of the memory management is to kep the cache for the
overlap (and derivative of overlap) to a manageable size so that memory doesn't
overflow. Since the value for each Slater determinant is stored, it is quite
feasible that the memory runs out during the optimization process

I thought it would be better to move the memory out of the wavfunction class
since it'll be harder to control the cache size without knowing the sizes of the
other objects. Also, caching is only relevant (I think) during the optimization
process which would suggest that it'd be better to move it closer to end of the workflow

1. Remove memory attribute from BaseWavefunction
2. Remove memory as parameter in __init__ of wavefunctions
3. Add method for loading cache in BaseSchrodinger
4. Move over tests