Remove cache system from BaseWavefunction
I think that as the cache system is implemented right now, it is a bit clunky
and makes it a little more difficult to implement new wavefunctions. Since the
goal is to make it as easy as possible to implement new wavefunctions, it would
be better to remove the cache system from the wavefunction altogether until a
better solution is available.

1. Remove load_cache method from BaseWavefunction
2. Remove clear_cache method from BaseWavefunction
3. Remove _cache_fns attributes from BaseWavefunction
4. Make corresponding changes to children, tests, docstrings.