Add caching system to the BaseSchrodinger
Since most of the information is available when constructing the Schrodinger
equation, we can have a pretty decent measure of the memory available for the
cache. Then, we can cache the overlap of the wavefunction (I think that it is
the only thing being cached). In order to avoid reconstructing the instance, we
overwrite the existing _olp and _olp_deriv method with their cached versions.

1. Add load_cache to BaseSchrodinger
2. Add clear_cache to BaseSchrodinger
3. Add tests
4. Change memory management code so that it is clear how much elements are
needed for the cache for _olp and for _olp_deriv. Also, if memory management is
changed if only one of the _olp and _olp_deriv is given/cached.

In order to distinguish the template _olp (raises NotImplementedError) and the
_olp with content, it is assumed that this method has an attribute called
"is_template", which gets removed when the _olp is overwritten