Remove NotImplementedError from _olp and _olp_deriv
Since caching is no longer incorporated directly inside the BaseWavefunction
instance, it is difficult to tell exactly when the _olp is used by the
get_overlap method. For example, in CIWavefunction, the _olp still raises the
NotImplementedError, but it isn't a problem because its get_overlap never calls
the cached function. In order to distinguish between the two, an attribute is
added to the function so that upon overwriting it, the attribute disappears.