Rename template_params to params_initial_guess
Before template_params functioned as the template of the parameters that is used
to check if the given parameters are good. However, now that there is a
params_shape and the dtype isn't rigorously monitored, it can be renamed for
what it actually is, an initial guess when the user does not provide one.

Though it is longer, it should be easier for people to understand