Remove old FIXME's
I think that the problem before was that the parameters of the schrodinger
equation were not separated according to the object to which they belong (i.e.
wavefunction vs hamiltonian). With the "new" ParamMask that selects the
appropriate parameters for the given objective, more sophisticated wavefunction
and hamiltonian structures should be supported.