Remove dtype reference in KerasNetwork.assign_params_initial_guess
Since dtype was removed as an attribute from BaseWavefunction, it is no longer
available to check the data type of the initial guess. However, since the
initial guess is fixed (by the method assign_params_initial_guess) it is not
even necessary to set the data type when constructing the initial guess (it is
float).