Remove wfns.ham.density_test
I think this module was original intended to be a refactor in progress of
wfns.ham.density. However, that was quite a long time ago, density module will
probably not get refactored in a long time. This module is also not tested, this
module does not belong in the package (right now anyways)