Fix bug in GeneralizedChemicalHamiltonian.assign_params
Background:
The derivative equations for the Hamiltonian integrals were derived such that
they are evaluated at zero. With this purpose in mind, the Hamiltonian
parameters are assigned and the integrals are updated/cached using the given
parameters. The parameters are incorporated into the integral by a forming a
unitary matrix and transforming the integrals.

During an optimization process, the derivatives are used to find the step vector
and this step vector is used to update the parameter. With subsequent calls to
the objective/jacobian, the parameters of the Hamiltonian is updated.

Problem:
1. The produced derivatives are assumed to be evaluated at zero. However, the
optimizer does not know this. It is given the full parameter of the
Hamiltonian (with respect to the initial integrals) rather than the relative one
after the integral update. Unless the Hamiltonian parameters are zero, optimizer
will use the wrong parameters.

2. The optimizer spits out the updated parameters which includes the step
vector. The step vector was produced with the assumption that the Hamiltonian
parameters are zero. Since the Hamiltonian parameters correspond to the unitary
transformation of the integrals, the transformation by the step vector must
correspond to a unique unitary matrix. In other words, the integrals must be
transformed by a sequence of unitary matrices that correspond to each step. In
general, the dot product of the unitary matrix of one step vector with the
unitary matrix of another step vector is not equal to the unitary matrix of the
sum of the two step vectors. Therefore, the sequence of unitary matrices must be
accounted for. Since the previous parameters are necessary to find the step
vector, the sequence of parameters must be accounted for also.